Critical update regarding NOTHING
We do not "affiliate" or promote contracts under the same umbrella as ones that don't work. This one in particular was a meme to promote other projects. Although I didn't submit for potential funding for helping, we cannot support projects that were hacked on the same day and experimental smart contracts.